code/admin-panel/requirements.txt
tkcalendar==1.6.1
mysql-connector-python==8.0.33
python-dotenv==1.0.0
